add movies event

+++ lib/movie/data/models/movie_model.dart
@@ -17,8 +17,7 @@ class MovieModel extends Movie {
         backdropPath: json['backdrop_path'],
         genreIds: List<int>.from(json['genre_ids'].map((e) => e)),
         overview: json['overview'],
-        // todo int is not subtype of double
-        voteAverage: json['vote_average'],
+        voteAverage: json['vote_average'].toDouble(),
         releaseDate: json['release_date'],
       );
 }
